Show the interface of the default route unless told to hide it.
Once settings are saved, KNemo only showed interfaces that were
explicitly configured.  Often this led to user frustration: an interface
that they were currently using was not visible.  To deal with this,
let's always display the interface of the default route unless we're
explicitly told to hide it.  That seems closer to the expected behavior.